% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_multi_series.R
\name{get_multi_series}
\alias{get_multi_series}
\title{Get Multi Series Data}
\usage{
get_multi_series(series_code, estimate)
}
\arguments{
\item{series_code:}{a vector with codes from series to fetch}

\item{estimate:}{logical to include or not projection from series}
}
\value{
a list with two dataframes, the first one is the series data and the second one is the series metadata.
}
\description{
Gets multiple series observations and metadata
}
\examples{
\dontrun{
    response <- get_multi_series(c("AREMP0085000OOQL", "BRBOP0044000OOML"), FALSE)
    response <- get_multi_series(c("AREMP0085000OOQL", "BRBOP0044000OOML"), TRUE)
 }
}
