% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_serie.R
\name{get_serie}
\alias{get_serie}
\title{Get Serie Data}
\usage{
get_serie(serie_code, estimate)
}
\arguments{
\item{serie_code:}{code from serie to fetch}

\item{estimate:}{logical to include or not projection from serie}
}
\value{
a list with two dataframes, the first one is the serie's data and the second one is the serie's metadata.
}
\description{
Gets serie observations and metadata
}
\examples{
\dontrun{
    response <- get_serie("AREMP0085000OOQL", FALSE)
    response <- get_serie("AREMP0085000OOQL", TRUE)
 }
}
