% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url.R
\name{get_url}
\alias{get_url}
\title{Getting url for modeling}
\usage{
get_url(endpoint, api)
}
\arguments{
\item{endpoint}{endpoint to access resource}
}
\value{
Api url.
}
\description{
Getting url for FaaS modeling.
}
\details{
DETAILS
}
\examples{
\dontrun{
url <- get_url("series")
}
}
